+++ CHANGELOG.md
@@ -114,7 +114,7 @@
 
 ## [0.7.4]
 
-### Changed
+### Merged from upstream
 
 - Added experimental [`zip_next::unstable::write::FileOptions::with_deprecated_encryption`] API to enable encrypting
   files with PKWARE encryption.